Fixes #7. BYU dropped the data explorer from production because since the update to version 1.1 of the NWM they haven’t addressed all of the changes with the file explorer. Another reason they removed the app from production is they are re-thinking a better way for users to access original NWM files as Tethys is not designed to be a file sharing system like IRODs-backed HydroShare or traditional FTP. Long time or intensive downloading consumes too much server resources that blocks other requests and even kills Tethys server.
The fix is to point pynwm to the development server, appsdev.hydroshare.org, on which the app is still running.  (I haven't tested to see how NWM 1.1 makes the app break.)

We should eventually rethink what role HydroShare has to play in pynwm. They have a new subsetting service that we aren't supporting yet but could be cool.  And the file downloads may go away altogether.